+++ app/Http/Controllers/Admin/SettingController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+
 use App\Http\Requests\SettingRequest;
 use App\Models\Setting;
 use Illuminate\Http\Request;